chore: [auto-generated] Update identification keys from upstream (#82)
Upstream Wynncraft API has been updated with new identification keys.

Please verify that the ids are correctly incremented, old ids are not
removed, and the new ids are incrementing from the last id.
If the ids reach 255, please reach out to the Artemis team to discuss
how to handle this. Most likely the gear chat encoding needs to have a
new version created.

This PR has been automatically generated.