feat: Switch back to using gear and gear_expanded jsons, move to using v3.3 item api (#152)
This PR is slowly deprecating the old, ...Advanced urls, and moving back
to the original urls. The advanced files have to be kept for a while, as
v3.3 item api introduces breaking format changes (due to 1.20.4+ model
changes), which we can only follow up on now. This means that Wynntils
2.0.0 will use the advanced routes, and Wynntils 2.1.0+ will use these,
original routes.