feat: Add crowd sourced services (#244)
This is likely incomplete as it's just from me but hopefully soon we can
get more people to collect anything I've missed and we'll finally have
an accurate map.

The data was collected whilst running
https://github.com/Wynntils/Wynntils/pull/3168 so until that is merged
into a release we can't collect any more.

3 fast travels were manually added due to a lack of fast travel label
being present for collection, those are:
Troms-Pigman ravine
Eltom-Thesead
Void Teleporter

I've made a python script to convert all the collected data into the
correct format, not sure if we want that on the repo or not

---------

Co-authored-by: ShadowCat117 <ShadowCat117@users.noreply.github.com>